+++ govcd/adminorg_test.go
@@ -227,7 +227,7 @@ func (vcd *TestVCD) TestOrg_AdminOrg_QueryCatalogList(check *C) {
 }
 
 // Test_GetAllVDCs checks that adminOrg.GetAllVDCs returns at least one VDC
-func (vcd *TestVCD) Test_GetAllVDCs(check *C) {
+func (vcd *TestVCD) Test_AdminOrgGetAllVDCs(check *C) {
 	if vcd.skipAdminTests {
 		check.Skip(fmt.Sprintf(TestRequiresSysAdminPrivileges, check.TestName()))
 	}
@@ -239,6 +239,26 @@ func (vcd *TestVCD) Test_GetAllVDCs(check *C) {
 	vdcs, err := adminOrg.GetAllVDCs(true)
 	check.Assert(err, IsNil)
 	check.Assert(len(vdcs) > 0, Equals, true)
+
+	// If NSX-T VDC is configured we expect to see at least 2 VDCs (NSX-V and NSX-T)
+	if vcd.config.VCD.Nsxt.Vdc != "" {
+		check.Assert(len(vdcs) >= 2, Equals, true)
+	}
+}
+
+func (vcd *TestVCD) Test_OrgGetAllVDCs(check *C) {
+	org, err := vcd.client.GetOrgByName(vcd.config.VCD.Org)
+	check.Assert(err, IsNil)
+	check.Assert(org, NotNil)
+
+	vdcs, err := org.GetAllVDCs(true)
+	check.Assert(err, IsNil)
+	check.Assert(len(vdcs) > 0, Equals, true)
+
+	// If NSX-T VDC is configured we expect to see at least 2 VDCs (NSX-V and NSX-T)
+	if vcd.config.VCD.Nsxt.Vdc != "" {
+		check.Assert(len(vdcs) >= 2, Equals, true)
+	}
 }
 
 // Test_GetAllStorageProfileReferences checks that adminOrg.GetAllStorageProfileReferences returns at least one storage